+++ CHANGELOG.rst
@@ -1,9 +1,8 @@
 new version (on deck)
 =====================
+- [ENH] Added function ``sort_timestamps_monotonically`` to timeseries functions @UGuntupalli
 - [ENH] Added the complete function for converting implicit missing values 
   to explicit ones. @samukweku
-
-
 v0.20.9
 =======
 - [ENH] Updated groupby_agg function to account for null entries in the ``by`` argument. @samukweku